+++ src/deps/vulkan/includelinux/vulkan/vk_enum_string_helper.h
@@ -7542,7 +7542,7 @@ static inline std::string string_VkToolPurposeFlags(VkToolPurposeFlags input_val
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkPipelineStageFlagBits2(uint64_t input_value) {
+static inline const char* string_VkPipelineStageFlagBits2(unsigned long long input_value) {
     if (input_value == VK_PIPELINE_STAGE_2_NONE) return "VK_PIPELINE_STAGE_2_NONE";
     if (input_value == VK_PIPELINE_STAGE_2_TOP_OF_PIPE_BIT) return "VK_PIPELINE_STAGE_2_TOP_OF_PIPE_BIT";
     if (input_value == VK_PIPELINE_STAGE_2_DRAW_INDIRECT_BIT) return "VK_PIPELINE_STAGE_2_DRAW_INDIRECT_BIT";
@@ -7596,7 +7596,7 @@ static inline std::string string_VkPipelineStageFlags2(VkPipelineStageFlags2 inp
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkPipelineStageFlagBits2(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkPipelineStageFlagBits2(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
@@ -7605,7 +7605,7 @@ static inline std::string string_VkPipelineStageFlags2(VkPipelineStageFlags2 inp
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkAccessFlagBits2(uint64_t input_value) {
+static inline const char* string_VkAccessFlagBits2(unsigned long long input_value) {
     if (input_value == VK_ACCESS_2_NONE) return "VK_ACCESS_2_NONE";
     if (input_value == VK_ACCESS_2_INDIRECT_COMMAND_READ_BIT) return "VK_ACCESS_2_INDIRECT_COMMAND_READ_BIT";
     if (input_value == VK_ACCESS_2_INDEX_READ_BIT) return "VK_ACCESS_2_INDEX_READ_BIT";
@@ -7661,7 +7661,7 @@ static inline std::string string_VkAccessFlags2(VkAccessFlags2 input_value) {
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkAccessFlagBits2(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkAccessFlagBits2(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
@@ -7728,7 +7728,7 @@ static inline std::string string_VkRenderingFlags(VkRenderingFlags input_value) 
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkFormatFeatureFlagBits2(uint64_t input_value) {
+static inline const char* string_VkFormatFeatureFlagBits2(unsigned long long input_value) {
     if (input_value == VK_FORMAT_FEATURE_2_SAMPLED_IMAGE_BIT) return "VK_FORMAT_FEATURE_2_SAMPLED_IMAGE_BIT";
     if (input_value == VK_FORMAT_FEATURE_2_STORAGE_IMAGE_BIT) return "VK_FORMAT_FEATURE_2_STORAGE_IMAGE_BIT";
     if (input_value == VK_FORMAT_FEATURE_2_STORAGE_IMAGE_ATOMIC_BIT) return "VK_FORMAT_FEATURE_2_STORAGE_IMAGE_ATOMIC_BIT";
@@ -7785,7 +7785,7 @@ static inline std::string string_VkFormatFeatureFlags2(VkFormatFeatureFlags2 inp
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkFormatFeatureFlagBits2(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkFormatFeatureFlagBits2(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
@@ -7819,7 +7819,7 @@ static inline std::string string_VkMemoryUnmapFlags(VkMemoryUnmapFlags input_val
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkPipelineCreateFlagBits2(uint64_t input_value) {
+static inline const char* string_VkPipelineCreateFlagBits2(unsigned long long input_value) {
     if (input_value == VK_PIPELINE_CREATE_2_DISABLE_OPTIMIZATION_BIT) return "VK_PIPELINE_CREATE_2_DISABLE_OPTIMIZATION_BIT";
     if (input_value == VK_PIPELINE_CREATE_2_ALLOW_DERIVATIVES_BIT) return "VK_PIPELINE_CREATE_2_ALLOW_DERIVATIVES_BIT";
     if (input_value == VK_PIPELINE_CREATE_2_DERIVATIVE_BIT) return "VK_PIPELINE_CREATE_2_DERIVATIVE_BIT";
@@ -7869,7 +7869,7 @@ static inline std::string string_VkPipelineCreateFlags2(VkPipelineCreateFlags2 i
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkPipelineCreateFlagBits2(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkPipelineCreateFlagBits2(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
@@ -7878,7 +7878,7 @@ static inline std::string string_VkPipelineCreateFlags2(VkPipelineCreateFlags2 i
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkBufferUsageFlagBits2(uint64_t input_value) {
+static inline const char* string_VkBufferUsageFlagBits2(unsigned long long input_value) {
     if (input_value == VK_BUFFER_USAGE_2_TRANSFER_SRC_BIT) return "VK_BUFFER_USAGE_2_TRANSFER_SRC_BIT";
     if (input_value == VK_BUFFER_USAGE_2_TRANSFER_DST_BIT) return "VK_BUFFER_USAGE_2_TRANSFER_DST_BIT";
     if (input_value == VK_BUFFER_USAGE_2_UNIFORM_TEXEL_BUFFER_BIT) return "VK_BUFFER_USAGE_2_UNIFORM_TEXEL_BUFFER_BIT";
@@ -7919,7 +7919,7 @@ static inline std::string string_VkBufferUsageFlags2(VkBufferUsageFlags2 input_v
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkBufferUsageFlagBits2(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkBufferUsageFlagBits2(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
@@ -9095,7 +9095,7 @@ static inline std::string string_VkVideoEncodeAV1RateControlFlagsKHR(VkVideoEnco
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkAccessFlagBits3KHR(uint64_t input_value) {
+static inline const char* string_VkAccessFlagBits3KHR(unsigned long long input_value) {
     if (input_value == VK_ACCESS_3_NONE_KHR) return "VK_ACCESS_3_NONE_KHR";
     return "Unhandled VkAccessFlagBits3KHR";
 }
@@ -9107,7 +9107,7 @@ static inline std::string string_VkAccessFlags3KHR(VkAccessFlags3KHR input_value
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkAccessFlagBits3KHR(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkAccessFlagBits3KHR(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
@@ -9911,7 +9911,7 @@ static inline std::string string_VkMicromapCreateFlagsEXT(VkMicromapCreateFlagsE
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkPhysicalDeviceSchedulingControlsFlagBitsARM(uint64_t input_value) {
+static inline const char* string_VkPhysicalDeviceSchedulingControlsFlagBitsARM(unsigned long long input_value) {
     if (input_value == VK_PHYSICAL_DEVICE_SCHEDULING_CONTROLS_SHADER_CORE_COUNT_ARM) return "VK_PHYSICAL_DEVICE_SCHEDULING_CONTROLS_SHADER_CORE_COUNT_ARM";
     return "Unhandled VkPhysicalDeviceSchedulingControlsFlagBitsARM";
 }
@@ -9923,7 +9923,7 @@ static inline std::string string_VkPhysicalDeviceSchedulingControlsFlagsARM(VkPh
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkPhysicalDeviceSchedulingControlsFlagBitsARM(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkPhysicalDeviceSchedulingControlsFlagBitsARM(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
@@ -9932,7 +9932,7 @@ static inline std::string string_VkPhysicalDeviceSchedulingControlsFlagsARM(VkPh
     return ret;
 }
 #endif // __cplusplus
-static inline const char* string_VkMemoryDecompressionMethodFlagBitsNV(uint64_t input_value) {
+static inline const char* string_VkMemoryDecompressionMethodFlagBitsNV(unsigned long long input_value) {
     if (input_value == VK_MEMORY_DECOMPRESSION_METHOD_GDEFLATE_1_0_BIT_NV) return "VK_MEMORY_DECOMPRESSION_METHOD_GDEFLATE_1_0_BIT_NV";
     return "Unhandled VkMemoryDecompressionMethodFlagBitsNV";
 }
@@ -9944,7 +9944,7 @@ static inline std::string string_VkMemoryDecompressionMethodFlagsNV(VkMemoryDeco
     while(input_value) {
         if (input_value & 1) {
             if( !ret.empty()) ret.append("|");
-            ret.append(string_VkMemoryDecompressionMethodFlagBitsNV(static_cast<uint64_t>(1ULL << index)));
+            ret.append(string_VkMemoryDecompressionMethodFlagBitsNV(static_cast<unsigned long long>(1ULL << index)));
         }
         ++index;
         input_value >>= 1;
